pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "valid8 (>=5.1.2,<6.0.0)",
    "typeguard (>=4.4.4,<5.0.0)",
    "coverage (>=7.12.0,<8.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "coreapi (>=2.3.3,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "drf-spectacular-sidecar (>=2025.12.1,<2026.0.0)",
    "django-stubs (>=5.2.8,<6.0.0)",
    "mixer (>=7.2.2,<8.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "dj-rest-auth (>=7.0.1,<8.0.0)",
    "django-allauth (>=65.13.1,<66.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]

# to exclude non-testable files
[tool.coverage.run]
source = ["."]
omit = [
    "*/tests/*",           # do not consider the tests themselves
    "*/migrations/*",
    "manage.py",
    "backend/asgi.py",
    "backend/wsgi.py"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
